Reorganization of tree, tests
This commit reorganizes the tree to make it easier to test individual
library functionality within it. Notably it:

- Moves everything under a library
- Moves zeronsd to the `src/bin` directory, with a small boot stub
- Moves the integration tests to `tests/`, with the service abstraction
  + library tooling for the tests in one spot.
- It also rewrites significant parts of the testing functionality to use
  tokio::test, with all the async functionality that needed to change to
  go along with that.
