+++ src/bin/zeronsd.rs
@@ -1,0 +1,5 @@
+use zeronsd::init::init;
+
+fn main() -> Result<(), anyhow::Error> {
+    init()
+}
