+++ src/lib.rs
@@ -1,0 +1,12 @@
+pub mod addresses;
+pub mod authority;
+pub mod cli;
+pub mod hosts;
+pub mod server;
+pub mod supervise;
+pub mod utils;
+
+pub mod init;
+
+#[cfg(test)]
+mod tests;
