+++ tests/integration.rs
@@ -1,50 +1,6 @@
-use trust_dns_resolver::{IntoName, Name};
+use zeronsd::addresses::Calculator;
 
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
-
-type SocketVec = Vec<SocketAddr>;
-
-pub(crate) trait ToIPv4Vec {
-    fn to_ipv4_vec(self) -> Vec<Ipv4Addr>;
-}
-
-pub(crate) trait ToIPv6Vec {
-    fn to_ipv6_vec(self) -> Vec<Ipv6Addr>;
-}
-
-pub(crate) trait ToPTRVec {
-    fn to_ptr_vec(self) -> Vec<Name>;
-}
-
-impl ToIPv4Vec for SocketVec {
-    fn to_ipv4_vec(self) -> Vec<Ipv4Addr> {
-        self.into_iter()
-            .filter_map(|ip| match ip.ip() {
-                IpAddr::V4(ip) => Some(ip),
-                IpAddr::V6(_) => None,
-            })
-            .collect::<Vec<Ipv4Addr>>()
-    }
-}
-
-impl ToIPv6Vec for SocketVec {
-    fn to_ipv6_vec(self) -> Vec<Ipv6Addr> {
-        self.into_iter()
-            .filter_map(|ip| match ip.ip() {
-                IpAddr::V4(_) => None,
-                IpAddr::V6(ip) => Some(ip),
-            })
-            .collect::<Vec<Ipv6Addr>>()
-    }
-}
-
-impl ToPTRVec for SocketVec {
-    fn to_ptr_vec(self) -> Vec<Name> {
-        self.into_iter()
-            .map(|ip| ip.ip().into_name().unwrap())
-            .collect::<Vec<Name>>()
-    }
-}
+mod service;
 
 mod sixplane {
     use std::{net::IpAddr, path::Path, str::FromStr, time::Duration};
@@ -53,21 +9,13 @@ mod sixplane {
     use rand::prelude::SliceRandom;
     use trust_dns_resolver::{IntoName, Name};
 
-    use crate::{
-        addresses::Calculator,
-        authority::{
-            service::{HostsType, Lookup, Service, ServiceConfig},
-            tests::ToIPv6Vec,
-        },
-        hosts::parse_hosts,
-        tests::HOSTS_DIR,
-        utils::init_logger,
-    };
+    use crate::service::{HostsType, Lookup, Service, ServiceConfig, ToIPv6Vec};
+    use zeronsd::{addresses::Calculator, hosts::parse_hosts, utils::init_logger};
 
-    #[test]
-    fn test_battery_single_domain() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain() {
         init_logger();
-        let service = Service::new(ServiceConfig::default().network_filename("6plane-only"));
+        let service = Service::new(ServiceConfig::default().network_filename("6plane-only")).await;
 
         let record = service.member_record();
 
@@ -76,25 +24,26 @@ mod sixplane {
         listen_ips.sort();
 
         for _ in 0..10000 {
-            let mut ips = service.lookup_aaaa(record.clone());
+            let mut ips = service.lookup_aaaa(record.clone()).await;
             ips.sort();
 
             assert_eq!(ips.sort(), listen_ips.clone().to_ipv6_vec().sort());
         }
     }
 
-    #[test]
-    fn test_battery_single_domain_named() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain_named() {
         init_logger();
         let update_interval = Duration::new(20, 0);
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(update_interval))
                 .network_filename("6plane-only"),
-        );
+        )
+        .await;
         let member_record = service.member_record();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         let named_record = "islay.domain.".to_string();
 
@@ -105,28 +54,31 @@ mod sixplane {
             listen_ips.sort();
 
             for _ in 0..10000 {
-                let mut ips = service.lookup_aaaa(record.clone());
+                let mut ips = service.lookup_aaaa(record.clone()).await;
                 ips.sort();
                 assert_eq!(ips, listen_ips.clone().to_ipv6_vec());
             }
         }
     }
 
-    #[test]
-    fn test_battery_multi_domain_hosts_file() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_multi_domain_hosts_file() {
         init_logger();
         let service = Service::new(
             ServiceConfig::default()
                 .hosts(HostsType::Fixture("basic-ipv6"))
                 .network_filename("6plane-only"),
-        );
+        )
+        .await;
 
         let record = service.member_record();
 
         info!("Looking up random domains");
 
         let mut hosts_map = parse_hosts(
-            Some(Path::new(&format!("{}/basic-ipv6", HOSTS_DIR)).to_path_buf()),
+            Some(
+                Path::new(&format!("{}/basic-ipv6", zeronsd::utils::TEST_HOSTS_DIR)).to_path_buf(),
+            ),
             "domain.".into_name().unwrap(),
         )
         .unwrap();
@@ -138,7 +90,7 @@ mod sixplane {
         for _ in 0..10000 {
             hosts.shuffle(&mut rand::thread_rng());
             let host = *hosts.first().unwrap();
-            let ip = service.lookup_aaaa(host.to_string());
+            let ip = service.lookup_aaaa(host.to_string()).await;
             assert!(hosts_map
                 .get(&IpAddr::V6(*ip.first().unwrap()))
                 .unwrap()
@@ -146,24 +98,26 @@ mod sixplane {
         }
     }
 
-    #[test]
-    fn test_wildcard_central() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_wildcard_central() {
         init_logger();
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(Duration::new(20, 0)))
                 .network_filename("6plane-only")
                 .wildcard_everything(true),
-        );
+        )
+        .await;
 
         let member_record = service.member_record();
         let named_record = Name::from_str("islay.domain.").unwrap();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         assert_eq!(
             service
                 .lookup_aaaa(named_record.to_string())
+                .await
                 .first()
                 .unwrap(),
             &service.clone().any_listen_ip()
@@ -172,6 +126,7 @@ mod sixplane {
         assert_eq!(
             service
                 .lookup_aaaa(member_record.to_string())
+                .await
                 .first()
                 .unwrap(),
             &service.clone().any_listen_ip()
@@ -184,7 +139,7 @@ mod sixplane {
                     .append_domain(&Name::from_str(&rec).unwrap())
                     .to_string();
                 assert_eq!(
-                    service.lookup_aaaa(lookup).first().unwrap(),
+                    service.lookup_aaaa(lookup).await.first().unwrap(),
                     &service.clone().any_listen_ip()
                 );
             }
@@ -199,21 +154,13 @@ mod rfc4193 {
     use rand::{prelude::SliceRandom, thread_rng};
     use trust_dns_resolver::{IntoName, Name};
 
-    use crate::{
-        addresses::Calculator,
-        authority::{
-            service::{HostsType, Lookup, Service, ServiceConfig},
-            tests::{ToIPv6Vec, ToPTRVec},
-        },
-        hosts::parse_hosts,
-        tests::HOSTS_DIR,
-        utils::init_logger,
-    };
+    use crate::service::{HostsType, Lookup, Service, ServiceConfig, ToIPv6Vec, ToPTRVec};
+    use zeronsd::{addresses::Calculator, hosts::parse_hosts, utils::init_logger};
 
-    #[test]
-    fn test_battery_single_domain() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain() {
         init_logger();
-        let service = Service::new(ServiceConfig::default().network_filename("rfc4193-only"));
+        let service = Service::new(ServiceConfig::default().network_filename("rfc4193-only")).await;
 
         let record = service.member_record();
 
@@ -222,17 +169,17 @@ mod rfc4193 {
         listen_ips.sort();
 
         for _ in 0..10000 {
-            let mut ips = service.lookup_aaaa(record.clone());
+            let mut ips = service.lookup_aaaa(record.clone()).await;
             ips.sort();
 
             assert_eq!(ips.sort(), listen_ips.clone().to_ipv6_vec().sort());
         }
 
-        let ptr_records: Vec<Name> = service
+        let ptr_records: Vec<String> = service
             .listen_ips
             .clone()
             .into_iter()
-            .map(|ip| ip.ip().into_name().unwrap())
+            .map(|ip| ip.ip().to_string())
             .collect();
 
         for ptr_record in ptr_records.clone() {
@@ -241,7 +188,11 @@ mod rfc4193 {
             for _ in 0..10000 {
                 let service = service.clone();
                 assert_eq!(
-                    service.lookup_ptr(ptr_record.to_string()).first().unwrap(),
+                    service
+                        .lookup_ptr(ptr_record.clone())
+                        .await
+                        .first()
+                        .unwrap(),
                     &record.to_string()
                 );
             }
@@ -253,7 +204,7 @@ mod rfc4193 {
             // randomly switch order
             if rand::random::<bool>() {
                 assert_eq!(
-                    service.lookup_aaaa(record.clone()).sort(),
+                    service.lookup_aaaa(record.clone()).await.sort(),
                     listen_ips.clone().to_ipv6_vec().sort()
                 );
 
@@ -261,6 +212,7 @@ mod rfc4193 {
                     service
                         .clone()
                         .lookup_ptr(ptr_records.choose(&mut thread_rng()).unwrap().to_string())
+                        .await
                         .first()
                         .unwrap(),
                     &record.to_string()
@@ -270,31 +222,34 @@ mod rfc4193 {
                     service
                         .clone()
                         .lookup_ptr(ptr_records.choose(&mut thread_rng()).unwrap().to_string())
+                        .await
                         .first()
                         .unwrap(),
                     &record.to_string()
                 );
 
                 assert_eq!(
-                    service.lookup_aaaa(record.clone()).sort(),
+                    service.lookup_aaaa(record.clone()).await.sort(),
                     listen_ips.clone().to_ipv6_vec().sort()
                 );
             }
         }
     }
 
-    #[test]
-    fn test_battery_single_domain_named() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain_named() {
         init_logger();
         let update_interval = Duration::new(20, 0);
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(update_interval))
                 .network_filename("rfc4193-only"),
-        );
+        )
+        .await;
+
         let member_record = service.member_record();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         let named_record = "islay.domain.".to_string();
 
@@ -305,13 +260,13 @@ mod rfc4193 {
             listen_ips.sort();
 
             for _ in 0..10000 {
-                let mut ips = service.lookup_aaaa(record.clone());
+                let mut ips = service.lookup_aaaa(record.clone()).await;
                 ips.sort();
                 assert_eq!(ips, listen_ips.clone().to_ipv6_vec());
             }
         }
 
-        let ptr_records: Vec<Name> = service.listen_ips.clone().to_ptr_vec();
+        let ptr_records: Vec<String> = service.listen_ips.clone().to_ptr_vec();
 
         for ptr_record in ptr_records {
             info!("Looking up {}", ptr_record);
@@ -319,28 +274,35 @@ mod rfc4193 {
             for _ in 0..10000 {
                 let service = service.clone();
                 assert_eq!(
-                    service.lookup_ptr(ptr_record.to_string()).first().unwrap(),
+                    service
+                        .lookup_ptr(ptr_record.clone())
+                        .await
+                        .first()
+                        .unwrap(),
                     &named_record.to_string()
                 );
             }
         }
     }
 
-    #[test]
-    fn test_battery_multi_domain_hosts_file() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_multi_domain_hosts_file() {
         init_logger();
         let service = Service::new(
             ServiceConfig::default()
                 .hosts(HostsType::Fixture("basic-ipv6"))
                 .network_filename("rfc4193-only"),
-        );
+        )
+        .await;
 
         let record = service.member_record();
 
         info!("Looking up random domains");
 
         let mut hosts_map = parse_hosts(
-            Some(Path::new(&format!("{}/basic-ipv6", HOSTS_DIR)).to_path_buf()),
+            Some(
+                Path::new(&format!("{}/basic-ipv6", zeronsd::utils::TEST_HOSTS_DIR)).to_path_buf(),
+            ),
             "domain.".into_name().unwrap(),
         )
         .unwrap();
@@ -352,7 +314,7 @@ mod rfc4193 {
         for _ in 0..10000 {
             hosts.shuffle(&mut rand::thread_rng());
             let host = *hosts.first().unwrap();
-            let ip = service.lookup_aaaa(host.to_string());
+            let ip = service.lookup_aaaa(host.to_string()).await;
             assert!(hosts_map
                 .get(&IpAddr::V6(*ip.first().unwrap()))
                 .unwrap()
@@ -360,24 +322,26 @@ mod rfc4193 {
         }
     }
 
-    #[test]
-    fn test_wildcard_central() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_wildcard_central() {
         init_logger();
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(Duration::new(20, 0)))
                 .network_filename("rfc4193-only")
                 .wildcard_everything(true),
-        );
+        )
+        .await;
 
         let member_record = service.member_record();
         let named_record = Name::from_str("islay.domain.").unwrap();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         assert_eq!(
             service
                 .lookup_aaaa(named_record.to_string())
+                .await
                 .first()
                 .unwrap(),
             &service.clone().any_listen_ip()
@@ -386,6 +350,7 @@ mod rfc4193 {
         assert_eq!(
             service
                 .lookup_aaaa(member_record.to_string())
+                .await
                 .first()
                 .unwrap(),
             &service.clone().any_listen_ip()
@@ -398,7 +363,7 @@ mod rfc4193 {
                     .append_domain(&Name::from_str(&rec).unwrap())
                     .to_string();
                 assert_eq!(
-                    service.lookup_aaaa(lookup).first().unwrap(),
+                    service.lookup_aaaa(lookup).await.first().unwrap(),
                     &service.clone().any_listen_ip()
                 );
             }
@@ -407,41 +372,46 @@ mod rfc4193 {
 }
 
 mod ipv4 {
-    use std::{str::FromStr, time::Duration};
+    use std::time::Duration;
 
     use log::info;
-    use rand::{prelude::SliceRandom, thread_rng};
-    use trust_dns_resolver::{IntoName, Name};
+    use std::str::FromStr;
+    use trust_dns_resolver::Name;
 
-    use crate::{
-        authority::{
-            service::{Lookup, Service, ServiceConfig},
-            tests::{ToIPv4Vec, ToPTRVec},
-        },
-        utils::init_logger,
-    };
+    use zeronsd::utils::init_logger;
 
-    #[test]
-    fn test_wildcard_central() {
+    use crate::service::{Lookup, Service, ServiceConfig, ToIPv4Vec, ToPTRVec};
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_wildcard_central() {
         init_logger();
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(Duration::new(20, 0)))
                 .wildcard_everything(true),
-        );
+        )
+        .await;
 
         let member_record = service.member_record();
         let named_record = Name::from_str("islay.domain.").unwrap();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         assert_eq!(
-            service.lookup_a(named_record.to_string()).first().unwrap(),
+            service
+                .lookup_a(named_record.to_string())
+                .await
+                .first()
+                .unwrap(),
             &service.clone().any_listen_ip()
         );
 
         assert_eq!(
-            service.lookup_a(member_record.to_string()).first().unwrap(),
+            service
+                .lookup_a(member_record.to_string())
+                .await
+                .first()
+                .unwrap(),
             &service.clone().any_listen_ip()
         );
 
@@ -452,21 +422,24 @@ mod ipv4 {
                     .append_domain(&Name::from_str(&rec).unwrap())
                     .to_string();
                 assert_eq!(
-                    service.lookup_a(lookup).first().unwrap(),
+                    service.lookup_a(lookup).await.first().unwrap(),
                     &service.clone().any_listen_ip()
                 );
             }
         }
     }
 
-    #[test]
-    fn test_battery_single_domain() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain() {
+        use rand::{seq::SliceRandom, thread_rng};
+
         init_logger();
         let service = Service::new(ServiceConfig::default().ips(Some(vec![
             "172.16.240.2",
             "172.16.240.3",
             "172.16.240.4",
-        ])));
+        ])))
+        .await;
 
         let record = service.member_record();
 
@@ -475,17 +448,17 @@ mod ipv4 {
         listen_ips.sort();
 
         for _ in 0..10000 {
-            let mut ips = service.lookup_a(record.clone());
+            let mut ips = service.lookup_a(record.clone()).await;
             ips.sort();
 
             assert_eq!(ips.sort(), listen_ips.clone().to_ipv4_vec().sort());
         }
 
-        let ptr_records: Vec<Name> = service
+        let ptr_records: Vec<String> = service
             .listen_ips
             .clone()
             .into_iter()
-            .map(|ip| ip.ip().into_name().unwrap())
+            .map(|ip| ip.ip().to_string())
             .collect();
 
         for ptr_record in ptr_records.clone() {
@@ -494,7 +467,11 @@ mod ipv4 {
             for _ in 0..10000 {
                 let service = service.clone();
                 assert_eq!(
-                    service.lookup_ptr(ptr_record.to_string()).first().unwrap(),
+                    service
+                        .lookup_ptr(ptr_record.clone())
+                        .await
+                        .first()
+                        .unwrap(),
                     &record.to_string()
                 );
             }
@@ -506,7 +483,7 @@ mod ipv4 {
             // randomly switch order
             if rand::random::<bool>() {
                 assert_eq!(
-                    service.lookup_a(record.clone()).sort(),
+                    service.lookup_a(record.clone()).await.sort(),
                     listen_ips.clone().to_ipv4_vec().sort()
                 );
 
@@ -514,6 +491,7 @@ mod ipv4 {
                     service
                         .clone()
                         .lookup_ptr(ptr_records.choose(&mut thread_rng()).unwrap().to_string())
+                        .await
                         .first()
                         .unwrap(),
                     &record.to_string()
@@ -523,31 +501,34 @@ mod ipv4 {
                     service
                         .clone()
                         .lookup_ptr(ptr_records.choose(&mut thread_rng()).unwrap().to_string())
+                        .await
                         .first()
                         .unwrap(),
                     &record.to_string()
                 );
 
                 assert_eq!(
-                    service.lookup_a(record.clone()).sort(),
+                    service.lookup_a(record.clone()).await.sort(),
                     listen_ips.clone().to_ipv4_vec().sort()
                 );
             }
         }
     }
 
-    #[test]
-    fn test_battery_single_domain_named() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_single_domain_named() {
         init_logger();
         let update_interval = Duration::new(20, 0);
         let service = Service::new(
             ServiceConfig::default()
                 .update_interval(Some(update_interval))
                 .ips(Some(vec!["172.16.240.2", "172.16.240.3", "172.16.240.4"])),
-        );
+        )
+        .await;
+
         let member_record = service.member_record();
 
-        service.change_name("islay");
+        service.change_name("islay").await;
 
         let named_record = "islay.domain.".to_string();
 
@@ -558,13 +539,13 @@ mod ipv4 {
             listen_ips.sort();
 
             for _ in 0..10000 {
-                let mut ips = service.lookup_a(record.clone());
+                let mut ips = service.lookup_a(record.clone()).await;
                 ips.sort();
                 assert_eq!(ips, listen_ips.clone().to_ipv4_vec());
             }
         }
 
-        let ptr_records: Vec<Name> = service.listen_ips.clone().to_ptr_vec();
+        let ptr_records: Vec<String> = service.listen_ips.clone().to_ptr_vec();
 
         for ptr_record in ptr_records {
             info!("Looking up {}", ptr_record);
@@ -572,7 +553,11 @@ mod ipv4 {
             for _ in 0..10000 {
                 let service = service.clone();
                 assert_eq!(
-                    service.lookup_ptr(ptr_record.to_string()).first().unwrap(),
+                    service
+                        .lookup_ptr(ptr_record.clone())
+                        .await
+                        .first()
+                        .unwrap(),
                     &named_record.to_string()
                 );
             }
@@ -585,12 +570,12 @@ mod all {
     use rand::prelude::SliceRandom;
     use trust_dns_resolver::{IntoName, Name};
 
-    use crate::{
-        authority::service::{HostsType, Lookup, Service, ServiceConfig},
+    use zeronsd::{
         hosts::parse_hosts,
-        tests::HOSTS_DIR,
-        utils::init_logger,
+        utils::{init_logger, TEST_HOSTS_DIR},
     };
+
+    use crate::service::{HostsType, Lookup, Service, ServiceConfig};
 
     use std::{
         net::{IpAddr, Ipv4Addr, Ipv6Addr},
@@ -600,22 +585,23 @@ mod all {
         time::Duration,
     };
 
-    #[test]
-    fn test_battery_multi_domain_hosts_file() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_battery_multi_domain_hosts_file() {
         init_logger();
         let ips = vec!["172.16.240.2", "172.16.240.3", "172.16.240.4"];
         let service = Service::new(
             ServiceConfig::default()
                 .hosts(HostsType::Fixture("basic"))
                 .ips(Some(ips.clone())),
-        );
+        )
+        .await;
 
         let record = service.member_record();
 
         info!("Looking up random domains");
 
         let mut hosts_map = parse_hosts(
-            Some(Path::new(&format!("{}/basic", HOSTS_DIR)).to_path_buf()),
+            Some(Path::new(&format!("{}/basic", TEST_HOSTS_DIR)).to_path_buf()),
             "domain.".into_name().unwrap(),
         )
         .unwrap();
@@ -631,7 +617,7 @@ mod all {
         for _ in 0..10000 {
             hosts.shuffle(&mut rand::thread_rng());
             let host = *hosts.first().unwrap();
-            let ips = service.lookup_a(host.to_string());
+            let ips = service.lookup_a(host.to_string()).await;
             assert!(hosts_map
                 .get(&IpAddr::from(*ips.first().unwrap()))
                 .unwrap()
@@ -639,8 +625,8 @@ mod all {
         }
     }
 
-    #[test]
-    fn test_hosts_file_reloading() {
+    #[tokio::test(flavor = "multi_thread")]
+    async fn test_hosts_file_reloading() {
         init_logger();
         let hosts_path = "/tmp/zeronsd-test-hosts";
         std::fs::write(hosts_path, "127.0.0.2 islay\n::2 islay\n").unwrap();
@@ -648,11 +634,13 @@ mod all {
             ServiceConfig::default()
                 .hosts(HostsType::Path(hosts_path))
                 .update_interval(Some(Duration::new(20, 0))),
-        );
+        )
+        .await;
 
         assert_eq!(
             service
                 .lookup_a("islay.domain.".to_string())
+                .await
                 .first()
                 .unwrap(),
             &Ipv4Addr::from_str("127.0.0.2").unwrap()
@@ -661,6 +649,7 @@ mod all {
         assert_eq!(
             service
                 .lookup_aaaa("islay.domain.".to_string())
+                .await
                 .first()
                 .unwrap(),
             &Ipv6Addr::from_str("::2").unwrap()
@@ -672,6 +661,7 @@ mod all {
         assert_eq!(
             service
                 .lookup_a("islay.domain.".to_string())
+                .await
                 .first()
                 .unwrap(),
             &Ipv4Addr::from_str("127.0.0.3").unwrap()
@@ -680,9 +670,70 @@ mod all {
         assert_eq!(
             service
                 .lookup_aaaa("islay.domain.".to_string())
+                .await
                 .first()
                 .unwrap(),
             &Ipv6Addr::from_str("::3").unwrap()
         );
     }
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_get_listen_ip() -> Result<(), anyhow::Error> {
+    use service::*;
+    use zeronsd::utils::*;
+
+    init_logger();
+
+    let tn = TestNetwork::new("basic-ipv4", &mut TestContext::default().await)
+        .await
+        .unwrap();
+
+    let listen_ips = get_listen_ips(&authtoken_path(None), &tn.network.clone().id.unwrap()).await?;
+
+    eprintln!("My listen IP is {}", listen_ips.first().unwrap());
+    assert_ne!(*listen_ips.first().unwrap(), String::from(""));
+
+    drop(tn);
+
+    // see testdata/networks/basic-ipv4.json
+    let mut ips = vec!["172.16.240.2", "172.16.240.3", "172.16.240.4"];
+    let tn =
+        TestNetwork::new_multi_ip("basic-ipv4", &mut TestContext::default().await, ips.clone())
+            .await
+            .unwrap();
+
+    let mut listen_ips =
+        get_listen_ips(&authtoken_path(None), &tn.network.clone().id.unwrap()).await?;
+
+    assert_eq!(listen_ips.sort(), ips.sort());
+    eprintln!("My listen IPs are {}", listen_ips.join(", "));
+
+    let tn = TestNetwork::new("rfc4193-only", &mut TestContext::default().await)
+        .await
+        .unwrap();
+
+    let mut listen_ips =
+        get_listen_ips(&authtoken_path(None), &tn.network.clone().id.unwrap()).await?;
+
+    let mut ips = vec![tn.member().clone().rfc4193()?.ip().to_string()];
+
+    assert_eq!(listen_ips.sort(), ips.sort());
+    eprintln!("My listen IPs are {}", listen_ips.join(", "));
+
+    drop(tn);
+
+    let tn = TestNetwork::new("6plane-only", &mut TestContext::default().await)
+        .await
+        .unwrap();
+
+    let mut listen_ips =
+        get_listen_ips(&authtoken_path(None), &tn.network.clone().id.unwrap()).await?;
+
+    let mut ips = vec![tn.member().clone().sixplane()?.ip().to_string()];
+
+    assert_eq!(listen_ips.sort(), ips.sort());
+    eprintln!("My listen IPs are {}", listen_ips.join(", "));
+
+    Ok(())
 }
